Issue #16: Added --dry-run and --*error to key commands

+++ cmd/key/delete.go
@@ -2,9 +2,12 @@ package key
 
 import (
 	"fmt"
+	"os"
 
+	"bitbucket.org/gildas_cherruel/bb/cmd/common"
 	"bitbucket.org/gildas_cherruel/bb/cmd/profile"
 	"bitbucket.org/gildas_cherruel/bb/cmd/user"
+	"github.com/gildas/go-errors"
 	"github.com/gildas/go-logger"
 	"github.com/spf13/cobra"
 )
@@ -19,13 +22,20 @@ var deleteCmd = &cobra.Command{
 }
 
 var deleteOptions struct {
-	Owner string
+	Owner        string
+	StopOnError  bool
+	WarnOnError  bool
+	IgnoreErrors bool
 }
 
 func init() {
 	Command.AddCommand(deleteCmd)
 
 	deleteCmd.Flags().StringVar(&deleteOptions.Owner, "user", "", "Owner of the keys")
+	deleteCmd.Flags().BoolVar(&deleteOptions.StopOnError, "stop-on-error", false, "Stop on error")
+	deleteCmd.Flags().BoolVar(&deleteOptions.WarnOnError, "warn-on-error", false, "Warn on error")
+	deleteCmd.Flags().BoolVar(&deleteOptions.IgnoreErrors, "ignore-errors", false, "Ignore errors")
+	deleteCmd.MarkFlagsMutuallyExclusive("stop-on-error", "warn-on-error", "ignore-errors")
 }
 
 func deleteValidArgs(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
@@ -48,19 +58,33 @@ func deleteProcess(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	log.Infof("Deleting GPG keys %v", args)
-
+	var merr errors.MultiError
 	for _, fingerprint := range args {
-		err := profile.Delete(
-			cmd.Context(),
-			cmd,
-			fmt.Sprintf("/users/%s/gpg-keys/%s", owner.ID, fingerprint),
-			nil,
-		)
-		if err != nil {
-			return err
+		if common.WhatIf(log.ToContext(cmd.Context()), cmd, "Deleting GPG key %s for user %s", fingerprint, owner.ID) {
+			err := profile.Delete(
+				cmd.Context(),
+				cmd,
+				fmt.Sprintf("/users/%s/gpg-keys/%s", owner.ID, fingerprint),
+				nil,
+			)
+			if err != nil {
+				if profile.ShouldStopOnError(cmd) {
+					fmt.Fprintf(os.Stderr, "Failed to delete key %s: %s\n", fingerprint, err)
+					os.Exit(1)
+				} else {
+					merr.Append(err)
+				}
+			}
+			log.Infof("Key %s deleted", fingerprint)
 		}
 	}
-
-	return nil
+	if !merr.IsEmpty() && profile.ShouldWarnOnError(cmd) {
+		fmt.Fprintf(os.Stderr, "Failed to delete these keys: %s\n", merr)
+		return nil
+	}
+	if profile.ShouldIgnoreErrors(cmd) {
+		log.Warnf("Failed to delete these keys, but ignoring errors: %s", merr)
+		return nil
+	}
+	return merr.AsError()
 }
